+++ src/common/CXMapPainter2D.cpp
@@ -1179,7 +1179,7 @@ void CXMapPainter2D::OnInternalPaint(IBitmap *pBMP, IBitmap *pTmpBMP, int Width,
         TextRect.OffsetRect(0, bottom);
         pBMP->DrawTextASCII(ttt, TextRect, MAPFGCOLOR, MAPBGCOLOR);
         bottom = TextRect.GetBottom();
-        int FreeMemMB = GetFreeMem()/1024/1024;
+        int FreeMemMB = static_cast<int>(GetFreeMem()/1024/1024);
         ttt = CXStringASCII("MB: ") + IToA<CXStringASCII>(FreeMemMB, 1);
         TextRect = pBMP->CalcTextRectASCII(ttt, 2, 2);
         TextRect.OffsetRect(0, bottom);
